add support for child objects, with the caveat that they add items on the path
