+++ index.js
@@ -123,7 +123,6 @@ export default function encrypt(db, keyOrPromise, cryptoSettings, onKeyChange, n
                 entity[rule.fields[i]] = hideValue(entity[rule.fields[i]]);
             }
         } else {
-            Object.assign(toEncrypt, entity);
             const indices = table.schema.indexes.map(index => index.name);
             const whitelist = rule.type === cryptoOptions.WHITELIST ? rule.fields : [];
             for (const key in entity) {
@@ -248,22 +247,10 @@ export default function encrypt(db, keyOrPromise, cryptoSettings, onKeyChange, n
                                                 Object.assign(obj, preservedValue);
                                             };
                                         });
-                                        table.hook('updating', function(
-                                            modifications,
-                                            primKey,
-                                            obj
-                                        ) {
-                                            const decrypted = decryptWithRule(
-                                                { ...obj },
-                                                newSetting
-                                            );
-                                            const updates = {
-                                                ...decrypted,
-                                                ...modifications,
-                                            };
+                                        table.hook('updating', function(modifications) {
                                             const encrypted = encryptWithRule(
                                                 table,
-                                                updates,
+                                                {...this.value},
                                                 newSetting
                                             );
                                             return encrypted;
